플로팅 버튼 1번째 화면 완료(토글 버튼 글자 색상 추후 수정)
