+++ planet-front/src/components/FloatingPart/Clock.js
@@ -1,0 +1,11 @@
+import React, { Component } from 'react';
+
+class Clock extends Component {
+    render() {
+      return (
+        <p>메롱</p>
+      );
+    }
+  }
+  
+  export default Clock;
