+++ planet-front/src/components/FloatingPart/Dashboard.js
@@ -5,29 +5,19 @@ import CustomSwitch from './CustomSwitch';
 //토글 스위치 Dashboard screen
 export default function Dashboard({navigation}){
     const onSelectSwitch = index => {
-      alert('Selected index: ' + index);
+      //수입, 지출화면으로 이동
     };
   
     return (
       <View style={{alignItems: 'center'}}>
-        <View style={{alignItems: 'center', margin: 20}}>
+        <View style={{alignItems: 'center', marginBottom: 60}}>
           <CustomSwitch
             selectionMode={1}
             roundCorner={true}
-            option1={'First'}
-            option2={'Second'}
+            option1={'수입'}
+            option2={'지출'}
             onSelectSwitch={onSelectSwitch}
-            selectionColor={'blue'}
-          />
-        </View>
-        <View style={{alignItems: 'center', margin: 20}}>
-          <CustomSwitch
-            selectionMode={2}
-            roundCorner={false}
-            option1={'Option1'}
-            option2={'Option2'}
-            onSelectSwitch={onSelectSwitch}
-            selectionColor={'red'}
+            selectionColor={'#374756'}
           />
         </View>
       </View>
