+++ planet-front/src/components/FloatingPart/TopNav.module.css
@@ -1,0 +1,28 @@
+@charset 'utf-8';
+@import url("https://cdn.jsdelivr.net/gh/orioncactus/pretendard/dist/web/static/pretendard.css");
+
+.backBtn {
+    position: relative;
+    width: 100%;
+    margin-left: 25px;
+    padding-top: 30px;
+    padding-bottom: 20px;
+    float: left;
+}
+
+.navBar {
+    position: absolute;
+    overflow: hidden;
+    margin: 0 auto;
+    text-align: center;
+    top: 13px;
+    left: 80px;
+}
+
+.navBar > li {
+    margin: 8px;
+    width: 10%;
+    float: left;
+    text-align: center;
+    list-style: none;
+}
